refactor(*-date-input): extract common styles
